Hard revert til pre-#210-tilstand
Tilbakefører alle filer berørt av #210-arbeidet til state på commit
5c7a80f (siste commit før PR #247).

Filer endret tilbake:
- app/(app)/chat/page.tsx
- app/(app)/profil/page.tsx
- components/TopHeader.tsx
- components/chat/Chat.tsx
- lib/actions/samtaler.ts

Filer slettet (lagt til av #210-arbeidet):
- components/chat/hooks/useChatData.ts
- components/chat/hooks/useChatScroll.ts
- components/chat/types.ts

Versjonsstamp + sw.js cache-bust beholdt fremover for å invalidere
klient-cache.
