+++ components/TopHeader.tsx
@@ -5,6 +5,7 @@ import { usePathname } from 'next/navigation'
 import { type CSSProperties, useEffect, useLayoutEffect, useRef, useState } from 'react'
 import Avatar from '@/components/ui/Avatar'
 import { harGulGloed } from '@/lib/roller'
+import { CHAT_TAB_PREFIKSER } from '@/lib/navigasjon'
 
 type Tab = {
   href: string
@@ -16,10 +17,7 @@ type Tab = {
 
 const TABS: Tab[] = [
   { href: '/', label: 'Agenda', nokkel: 'agenda', prefikser: ['/poll', '/arrangementer', '/meldinger'] },
-  // /samtaler aktiverer IKKE chat-tabben visuelt. Privatmeldinger åpnes fra
-  // profil-siden (#210 PR 4). Pull-to-refresh-deaktivering styres fortsatt
-  // av CHAT_TAB_PREFIKSER i lib/navigasjon.ts som beholder begge.
-  { href: '/chat', label: 'Chat', nokkel: 'chat', prefikser: ['/chat'] },
+  { href: '/chat', label: 'Chat', nokkel: 'chat', prefikser: [...CHAT_TAB_PREFIKSER] },
   { href: '/klubbinfo', label: 'Klubb', nokkel: 'klubb', prefikser: ['/klubbinfo', '/kaaringer', '/album'] },
 ]
