+++ lib/versjon.json
@@ -1,4 +1,4 @@
 {
-  "nummer": 51,
-  "versjon": "V3.1.51"
+  "nummer": 52,
+  "versjon": "V3.1.52"
 }
